+++ src/hipay_enterprise/classes/helper/HipayHelper.php
@@ -715,7 +715,6 @@ class HipayHelper
         $duplicationCart = $cart->duplicate();
 
         foreach ($cartRules as $rule) {
-            $duplicationCart['cart']->addCartRule($rule['id_cart_rule']);
             // If the discount is a gift, you don't want to re-apply the discount to add another quantity.
             if (!empty($rule['gift_product']) && (int) $rule['gift_product'] > 0) {
                 $duplicationCart['cart']->updateQty(-1, $rule['gift_product'], $rule['gift_product_attribute']);
@@ -1132,5 +1131,4 @@ class HipayHelper
             return array_intersect_key($card, array_flip($fieldsToKeep));
         }, $cards);
     }
-
 }